Fixed bug, added reusable Log method and cleaned the code with the extra time I had
I wanted to complete this as I felt the hour was not enough time for me to fully demonstrate my abilities. Understandably this branch should not be used as part of the code test as it was created and worked on outside of the hour provided for the challenge.

I fixed the bug with the loop along with adding a Log method to avoid having plain write lines everywhere, I also moved and tidied the code alittle along with fiddling with the maths within the loops.